+++ src/Command/Install/CreateGroupsCommand.php
@@ -19,7 +19,6 @@ use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
 use Cake\Database\Driver\Postgres;
 use Cake\Database\Driver\Sqlite;
-use Cake\Datasource\ConnectionManager;
 use MeTools\Console\Command;
 
 /**
@@ -27,6 +26,16 @@ use MeTools\Console\Command;
  */
 class CreateGroupsCommand extends Command
 {
+    /**
+     * Hook method invoked by CakePHP when a command is about to be executed
+     * @return void
+     */
+    public function initialize(): void
+    {
+        parent::initialize();
+        $this->loadModel('MeCms.UsersGroups');
+    }
+
     /**
      * Hook method for defining this command's option parser
      * @param \Cake\Console\ConsoleOptionParser $parser The parser to be defined
@@ -45,14 +54,12 @@ class CreateGroupsCommand extends Command
      */
     public function execute(Arguments $args, ConsoleIo $io): ?int
     {
-        $this->loadModel('MeCms.UsersGroups');
-
         if (!$this->UsersGroups->find()->isEmpty()) {
             return $io->error(__d('me_cms', 'Some user groups already exist'));
         }
 
         //Truncates the table (this resets IDs), then saves groups
-        $connection = ConnectionManager::get('default');
+        $connection = $this->UsersGroups->getConnection();
         $command = 'TRUNCATE TABLE `%s`';
         if ($connection->getDriver() instanceof Sqlite) {
             $command = 'DELETE FROM "sqlite_sequence" WHERE "name"=\'%s\';';