+++ tests/TestCase/Command/Install/CreateGroupsCommandTest.php
@@ -13,7 +13,13 @@ declare(strict_types=1);
  */
 namespace MeCms\Test\TestCase\Command\Install;
 
-use Cake\ORM\TableRegistry;
+use Cake\Console\Arguments;
+use Cake\Console\ConsoleIo;
+use Cake\Database\Connection;
+use Cake\Database\Driver\Postgres;
+use Cake\Database\Driver\Sqlite;
+use Cake\ORM\Query;
+use Cake\ORM\Table;
 use MeCms\TestSuite\TestCase;
 use MeTools\TestSuite\ConsoleIntegrationTestTrait;
 
@@ -23,6 +29,11 @@ use MeTools\TestSuite\ConsoleIntegrationTestTrait;
 class CreateGroupsCommandTest extends TestCase
 {
     use ConsoleIntegrationTestTrait;
+
+    /**
+     * @var bool
+     */
+    public $autoInitializeClass = true;
 
     /**
      * Fixtures
@@ -39,8 +50,6 @@ class CreateGroupsCommandTest extends TestCase
      */
     public function testExecute()
     {
-        $UsersGroups = TableRegistry::getTableLocator()->get('MeCms.UsersGroups');
-
         //A group already exists
         $this->exec('me_cms.create_groups -v');
         $this->assertExitWithSuccess();
@@ -48,6 +57,7 @@ class CreateGroupsCommandTest extends TestCase
         $this->assertErrorContains('Some user groups already exist');
 
         //With no user groups
+        $UsersGroups = $this->getTable('MeCms.UsersGroups');
         $UsersGroups->deleteAll(['id is NOT' => null]);
         $this->exec('me_cms.create_groups -v');
         $this->assertExitWithSuccess();
@@ -56,5 +66,53 @@ class CreateGroupsCommandTest extends TestCase
 
         //Checks the user groups exist
         $this->assertEquals([1, 2, 3], $UsersGroups->find()->extract('id')->toList());
+    }
+
+    /**
+     * Provider for `testExecuteOtherDrivers()`
+     * @return array
+     */
+    public function driverProvider(): array
+    {
+        return [
+            'postgres' => [Postgres::class],
+            'sqlite' => [Sqlite::class],
+        ];
+    }
+
+    /**
+     * Test for `execute()` method
+     * @dataProvider driverProvider
+     * @test
+     */
+    public function testExecuteOtherDrivers($driver)
+    {
+        $this->skipIf(IS_WIN);
+
+        $this->Command->UsersGroups = $this->getMockBuilder(Table::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->Command->UsersGroups->method('find')->will($this->returnCallback(function () {
+            $query = $this->getMockBuilder(Query::class)
+                ->disableOriginalConstructor()
+                ->setMethods(array_merge(get_class_methods(Query::class), ['isEmpty']))
+                ->getMock();
+            $query->method('isEmpty')->will($this->returnValue(true));
+
+            return $query;
+        }));
+
+        $this->Command->UsersGroups->method('getConnection')->will($this->returnCallback(function () use ($driver) {
+            $driver = $this->getMockBuilder($driver)->getMock();
+            $driver->method('enabled')->will($this->returnValue(true));
+
+            return $this->getMockBuilder(Connection::class)
+                ->setConstructorArgs([compact('driver')])
+                ->setMethods(['execute'])
+                ->getMock();
+        }));
+
+        $this->assertNull($this->Command->execute(new Arguments([], [], [], []), new ConsoleIo()));
     }
 }