+++ tests/bootstrap.php
@@ -101,12 +101,13 @@ Cache::setConfig([
 //Ensure default test connection is defined
 if (!getenv('db_dsn')) {
     putenv('db_dsn=mysql://travis@localhost/test');
-}
-if (!getenv('db_dsn_postgres')) {
-    putenv('db_dsn_postgres=postgres://postgres@localhost/travis_ci_test');
+
+    if (getenv('db_driver') == 'postgres') {
+        putenv('db_dsn=postgres://postgres@localhost/travis_ci_test');
+    }
 }
 ConnectionManager::setConfig('test', ['url' => getenv('db_dsn')]);
-ConnectionManager::setConfig('test_postgres', ['url' => getenv('db_dsn_postgres')]);
+echo 'Running tests for "' . ConnectionManager::getConfig('test')['scheme'] . '" driver ' . PHP_EOL;
 
 //This adds `apache_get_modules()` and `apache_get_version()` functions
 require_once VENDOR . 'mirko-pagliai' . DS . 'php-tools' . DS . 'tests' . DS . 'apache_functions.php';