+++ tests/test_app/TestApp/Plugin/DebugKit/src/Plugin.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace DebugKit;
+
+use Cake\Core\BasePlugin;
+
+class Plugin extends BasePlugin
+{
+
+}